fixed gcc compile warningas
<==== Mercurial data ====>
Branch: release_2.3_review

+++ unit_tests/common/src/utils/network/socket_test_helper.cpp
@@ -311,7 +311,7 @@ void ConnectionsGenerator::join()
 
 bool ConnectionsGenerator::start()
 {
-    for( int i = 0; i < m_maxSimultaneousConnectionsCount; ++i )
+    for( size_t i = 0; i < m_maxSimultaneousConnectionsCount; ++i )
     {
         std::unique_lock<std::mutex> lk( m_mutex );
